requirements.txt
celery==4.2.1
django==2.1.1
djangorestframework==3.8.2
psycopg2-binary==2.7.5
redis==2.10.6
requests==2.19.1
